NVML Support
Either NVML or ADL may be active depending on a platform vendor.
NVML is run time linked to libnvidia-ml.so (Linux) or nvml.dll (Windows)
which are supposed to be found in a system path or miner's directory.
NVML provides the miner with temperature and fan speed data.

nvml API v6

tpruvot: Modified to be functional :p